mymcpserver/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fitness-coach-mcp"
version = "0.1.0"
description = "MCP server combining Strava + Garmin fitness data"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.0.0",
    "stravalib>=1.0.0",
    "garminconnect>=0.2.38",
    "python-dotenv>=1.0.0",
]
